+++ gui/src/app/pages/NwbPage/viewPlugins/TimeIntervals/NwbTimeIntervalsView.tsx
@@ -21,10 +21,65 @@ const NwbTimeIntervalsView: FunctionComponent<Props> = ({
   const { data: startTimeData } = useDatasetData(nwbFile, `${path}/start_time`);
   const { data: stopTimeData } = useDatasetData(nwbFile, `${path}/stop_time`);
 
-  // auto detect label data
+  const { labelFieldName, labelData } = useLabelData(
+    nwbFile,
+    path,
+    startTimeData?.length,
+  );
+
+  if (!startTimeData || !stopTimeData) {
+    return <div>loading data...</div>;
+  }
+
+  const bottomBarHeight = 50;
+
+  return (
+    <div style={{ position: "relative", width, height }}>
+      <div
+        style={{
+          position: "absolute",
+          width,
+          height: height - bottomBarHeight,
+          top: 0,
+        }}
+      >
+        <NwbTimeIntervalsWidget
+          labels={labelData}
+          startTimes={startTimeData}
+          stopTimes={stopTimeData}
+          width={width}
+          height={height - bottomBarHeight}
+        />
+      </div>
+      <div
+        style={{
+          position: "absolute",
+          width,
+          height: bottomBarHeight,
+          top: height - bottomBarHeight,
+        }}
+      >
+        {labelFieldName && (
+          <div style={{ paddingLeft: 10 }}>Label: {labelFieldName}</div>
+        )}
+      </div>
+    </div>
+  );
+};
+
+const useLabelData = (
+  nwbFile: any,
+  path: string,
+  numRows: number | undefined,
+) => {
+  const group = useGroup(nwbFile, path);
+  const [labelFieldName, setLabelFieldName] = useState<string | undefined>(
+    undefined,
+  );
   const [labelData, setLabelData] = useState<string[] | undefined>(undefined);
   useEffect(() => {
     let canceled = false;
+    setLabelFieldName(undefined);
     setLabelData(undefined);
     const load = async () => {
       if (!group) return;
@@ -37,12 +92,13 @@ const NwbTimeIntervalsView: FunctionComponent<Props> = ({
         if (canceled) return;
         try {
           const values = Array.from(d) as any as string[];
+          if (values.length !== numRows) continue; // restrict to columns with same number of rows as time columns
           const distinctValues = getDistinctValues(values);
           if (
             distinctValues.length > 1 &&
             distinctValues.length <= values.length / 4
           ) {
-            const score = distinctValues.length / values.length;
+            const score = Math.abs(distinctValues.length - 10); // we target 10 distinct values
             scores.push({ colname, score, values });
           }
         } catch (err) {
@@ -51,6 +107,7 @@ const NwbTimeIntervalsView: FunctionComponent<Props> = ({
       }
       scores.sort((a, b) => a.score - b.score);
       if (scores.length > 0) {
+        setLabelFieldName(scores[0].colname);
         setLabelData(scores[0].values);
       }
     };
@@ -58,21 +115,8 @@ const NwbTimeIntervalsView: FunctionComponent<Props> = ({
     return () => {
       canceled = true;
     };
-  }, [group, nwbFile, path]);
-
-  if (!startTimeData || !stopTimeData) {
-    return <div>loading data...</div>;
-  }
-
-  return (
-    <NwbTimeIntervalsWidget
-      labels={labelData}
-      startTimes={startTimeData}
-      stopTimes={stopTimeData}
-      width={width}
-      height={height}
-    />
-  );
+  }, [group, nwbFile, path, numRows]);
+  return { labelFieldName, labelData };
 };
 
 const getDistinctValues = (values: string[]) => {